Changes the modem firmware download process to work on the SGS devices, fixed the GPIO mapping. Use the right GPIO to tell the BP it should turn on. The SGS BP Bootloader won't tell us that is has booted properly. We miss the magic value MODEM_MSG_SBL_DONE. So we wait some time an then continue booting the BP.
The firmware is send from userspace in libsec-ril.so what is loaded by the rild.
It reads the firmware from /dev/mtd/mtd5ro and passes it to the modem. It is tested with
JPM modem from GT-I9000. Userspace won't notice that the radio interface is there until you lock the phone
and unlock it. Looks like a little phone state polling problem.